Fix some issues with writer
